+++ Twitter+OAuth/MGTwitterEngine/MGTwitterEngine.h
@@ -63,7 +63,7 @@
 #endif
 
 // Connection methods
-- (int)numberOfConnections;
+- (NSUInteger)numberOfConnections;
 - (NSArray *)connectionIdentifiers;
 - (void)closeConnection:(NSString *)identifier;
 - (void)closeAllConnections;
